Add Playwright tests for rich text
